revamped SentinelImportActor into bidirectional SentinelConnectorActor, which can now also send commands to the Delphire server
added a "diagnostics" panel to the SentinelRoute UI, which is used to send commands and display responses. Only function so far is a inject command that uploads a fire image to the selected device

added SentinelCommand, extended SentinelNotification. Both event and action notification will still change to use a regular payload syntax. SentinelCommand still needs a requestId that is sent to and subsequently received from the Delphire server to identify which browser the response should be sent to